pyproject.toml
[tool.poetry]
name = "canalservisproject"
version = "0.1.0"
description = ""
authors = ["arturkalimullin <noti1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.4.0"
gunicorn = "^20.1.0"
redis = "^4.5.1"
celery = "^5.2.7"
google-api-python-client = "^2.81.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
lxml = "^4.9.2"
oauth2client = "^4.1.3"
httplib2 = "^0.21.0"
django-money = "^3.0.0"
aiogram = "^2.25.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
